@@ -1,12 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
+import { ProductCatalogRoutingModule } from './product-catalog-routing.module';
+import { ProductCatalogComponent } from './pages/catalog.component';
 
 
 @NgModule({
-  declarations: [],
+  declarations: [
+    ProductCatalogComponent
+  ],
   imports: [
-    CommonModule
+    CommonModule,
+    ProductCatalogRoutingModule
   ]
 })
 export class ProductCatalogModule { }
