@@ -1,12 +1,16 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
+import { ShoppingCartRoutingModule } from './shopping-cart-routing.module';
+import { ShoppingcartComponent } from './pages/shoppingcart.component';
 
 
 @NgModule({
-  declarations: [],
+  declarations: [
+    ShoppingcartComponent
+  ],
   imports: [
-    CommonModule
+    CommonModule,
+    ShoppingCartRoutingModule
   ]
 })
 export class ShoppingCartModule { }
